+++ src/SWP/Bundle/BridgeBundle/spec/Client/GuzzleApiClientSpec.php
@@ -8,9 +8,10 @@
  * For the full copyright and license information, please see the
  * AUTHORS and LICENSE files distributed with this source code.
  *
- * @copyright 2015 Sourcefabric z.ú.
+ * @copyright 2015 Sourcefabric z.ú
  * @license http://www.superdesk.org/license
  */
+
 namespace spec\SWP\Bundle\BridgeBundle\Client;
 
 use PhpSpec\ObjectBehavior;