+++ src/SWP/Bundle/BridgeBundle/spec/Doctrine/ORM/PackageRepositorySpec.php
@@ -8,9 +8,10 @@
  * For the full copyright and license information, please see the
  * AUTHORS and LICENSE files distributed with this source code.
  *
- * @copyright 2016 Sourcefabric z.ú.
+ * @copyright 2016 Sourcefabric z.ú
  * @license http://www.superdesk.org/license
  */
+
 namespace spec\SWP\Bundle\BridgeBundle\Doctrine\ORM;
 
 use Doctrine\ORM\EntityManagerInterface;