+++ src/SWP/Bundle/ContentBundle/Doctrine/ODM/PHPCR/RouteObjectInterface.php
@@ -8,28 +8,26 @@
  * For the full copyright and license information, please see the
  * AUTHORS and LICENSE files distributed with this source code.
  *
- * @copyright 2016 Sourcefabric z.ú.
+ * @copyright 2016 Sourcefabric z.ú
  * @license http://www.superdesk.org/license
  */
+
 namespace SWP\Bundle\ContentBundle\Doctrine\ODM\PHPCR;
 
 use Doctrine\ODM\PHPCR\HierarchyInterface;
 use SWP\Bundle\ContentBundle\Model\RouteInterface;
+use Symfony\Cmf\Component\Routing\RouteObjectInterface as BaseRouteObjectInterface;
 
-interface RouteObjectInterface extends RouteInterface, HierarchyInterface
+/**
+ * Interface RouteObjectInterface.
+ */
+interface RouteObjectInterface extends RouteInterface, HierarchyInterface, BaseRouteObjectInterface
 {
     /**
      * Set the object this url points to.
      *
      * @param mixed $object A content object that can be persisted by the
-     *                      storage layer.
+     *                      storage layer
      */
     public function setContent($object);
-
-    /**
-     * Get the object this url points to.
-     *
-     * @return string
-     */
-    public function getContent();
 }