+++ src/SWP/Bundle/CoreBundle/Detection/DeviceDetection.php
@@ -11,9 +11,10 @@
  * Some parts of that file were taken from the Liip/ThemeBundle
  * (c) Liip AG
  *
- * @copyright 2016 Sourcefabric z.ú.
+ * @copyright 2016 Sourcefabric z.ú
  * @license http://www.superdesk.org/license
  */
+
 namespace SWP\Bundle\CoreBundle\Detection;
 
 class DeviceDetection implements DeviceDetectionInterface