+++ src/SWP/Bundle/CoreBundle/Enhancer/RouteEnhancer.php
@@ -8,9 +8,10 @@
  * For the full copyright and license information, please see the
  * AUTHORS and LICENSE files distributed with this source code.
  *
- * @copyright 2016 Sourcefabric z.ú.
+ * @copyright 2016 Sourcefabric z.ú
  * @license http://www.superdesk.org/license
  */
+
 namespace SWP\Bundle\CoreBundle\Enhancer;
 
 use SWP\Component\TemplatesSystem\Gimme\Context\Context;
@@ -19,7 +20,7 @@ use Symfony\Component\HttpFoundation\Request;
 use SWP\Component\TemplatesSystem\Gimme\Loader\LoaderInterface;
 use SWP\Bundle\CoreBundle\Resolver\TemplateNameResolverInterface;
 use SWP\Bundle\ContentBundle\Model\ArticleInterface;
-use Symfony\Cmf\Component\Routing\RouteObjectInterface;
+use SWP\Bundle\ContentBundle\Doctrine\ODM\PHPCR\RouteObjectInterface;
 use SWP\Bundle\ContentBundle\Model\RouteInterface;
 use SWP\Bundle\CoreBundle\Controller\ContentController;