+++ src/SWP/Bundle/CoreBundle/Theme/Model/ThemeInterface.php
@@ -8,9 +8,10 @@
  * For the full copyright and license information, please see the
  * AUTHORS and LICENSE files distributed with this source code.
  *
- * @copyright 2016 Sourcefabric z.ú.
+ * @copyright 2016 Sourcefabric z.ú
  * @license http://www.superdesk.org/license
  */
+
 namespace SWP\Bundle\CoreBundle\Theme\Model;
 
 use Sylius\Bundle\ThemeBundle\Model\ThemeInterface as BaseThemeInterface;