+++ src/SWP/Bundle/FixturesBundle/DataFixtures/PHPCR/LoadArticlesData.php
@@ -8,9 +8,10 @@
  * For the full copyright and license information, please see the
  * AUTHORS and LICENSE files distributed with this source code.
  *
- * @copyright 2015 Sourcefabric z.ú.
+ * @copyright 2015 Sourcefabric z.ú
  * @license http://www.superdesk.org/license
  */
+
 namespace SWP\Bundle\FixturesBundle\DataFixtures\PHPCR;
 
 use Doctrine\Common\DataFixtures\FixtureInterface;
@@ -61,11 +62,20 @@ class LoadArticlesData extends AbstractFixture implements FixtureInterface, Orde
                         'slug' => null,
                     ],
                     'templateName' => 'news.html.twig',
+                    'articlesTemplateName' => 'article.html.twig',
                 ],
                 [
                     'parent' => $this->defaultTenantPrefix.'/routes',
                     'name' => 'articles',
-                    'type' => 'content',
+                    'type' => 'collection',
+                    'variablePattern' => '/{slug}',
+                    'requirements' => [
+                        'slug' => '[a-zA-Z1-9\-_\/]+',
+                    ],
+                    'type' => 'collection',
+                    'defaults' => [
+                        'slug' => null,
+                    ],
                 ],
                 [
                     'parent' => $this->defaultTenantPrefix.'/routes/articles',
@@ -143,6 +153,10 @@ class LoadArticlesData extends AbstractFixture implements FixtureInterface, Orde
 
             if (isset($routeData['templateName'])) {
                 $route->setTemplateName($routeData['templateName']);
+            }
+
+            if (isset($routeData['articlesTemplateName'])) {
+                $route->setArticlesTemplateName($routeData['articlesTemplateName']);
             }
 
             if (isset($routeData['defaults'])) {