+++ src/SWP/Component/Common/spec/Generator/RandomStringGeneratorSpec.php
@@ -8,9 +8,10 @@
  * For the full copyright and license information, please see the
  * AUTHORS and LICENSE files distributed with this source code.
  *
- * @copyright 2015 Sourcefabric z.ú.
+ * @copyright 2015 Sourcefabric z.ú
  * @license http://www.superdesk.org/license
  */
+
 namespace spec\SWP\Component\Common\Generator;
 
 use PhpSpec\ObjectBehavior;
@@ -22,23 +23,23 @@ use SWP\Component\Common\Generator\RandomStringGenerator;
  */
 class RandomStringGeneratorSpec extends ObjectBehavior
 {
-    function it_is_initializable()
+    public function it_is_initializable()
     {
         $this->shouldHaveType(RandomStringGenerator::class);
     }
 
-    function it_should_implement_interface()
+    public function it_should_implement_interface()
     {
         $this->shouldImplement(GeneratorInterface::class);
     }
 
-    function it_should_generate_code()
+    public function it_should_generate_code()
     {
         $this->generate(4)->shouldMatch('/^[a-z0-9]+$/');
         $this->generate(4)->shouldHaveLength(4);
     }
 
-    function it_should_throw_exception_when_empty_length()
+    public function it_should_throw_exception_when_empty_length()
     {
         $this->shouldThrow(\InvalidArgumentException::class)
             ->during('generate', [null]);