+++ src/SWP/Component/MultiTenancy/spec/Factory/OrganizationFactorySpec.php
@@ -8,9 +8,10 @@
  * For the full copyright and license information, please see the
  * AUTHORS and LICENSE files distributed with this source code.
  *
- * @copyright 2016 Sourcefabric z.ú.
+ * @copyright 2016 Sourcefabric z.ú
  * @license http://www.superdesk.org/license
  */
+
 namespace spec\SWP\Component\MultiTenancy\Factory;
 
 use PhpSpec\ObjectBehavior;
@@ -25,30 +26,30 @@ use SWP\Component\Storage\Factory\FactoryInterface;
  */
 class OrganizationFactorySpec extends ObjectBehavior
 {
-    function let(
+    public function let(
         FactoryInterface $factory,
         GeneratorInterface $generator
     ) {
         $this->beConstructedWith($factory, $generator);
     }
 
-    function it_is_initializable()
+    public function it_is_initializable()
     {
         $this->shouldHaveType(OrganizationFactory::class);
     }
 
-    function it_implements_organization_factory_interface()
+    public function it_implements_organization_factory_interface()
     {
         $this->shouldImplement(OrganizationFactoryInterface::class);
     }
 
-    function it_creates_empty_organization(FactoryInterface $factory, OrganizationInterface $organization)
+    public function it_creates_empty_organization(FactoryInterface $factory, OrganizationInterface $organization)
     {
         $factory->create()->willReturn($organization);
         $this->create()->shouldReturn($organization);
     }
 
-    function it_creates_a_new_organization_with_code(
+    public function it_creates_a_new_organization_with_code(
         FactoryInterface $factory,
         GeneratorInterface $generator,
         OrganizationInterface $organization