+++ CHANGELOG.md
@@ -9,6 +9,7 @@ To get the diff for a specific change, go to https://github.com/superdesk/web-pu
 
 * 0.1.0-5 
 
+ * bug [#129] [CoreBundle][ContentBundle] Change template name discovery, add articles_template_name to route (see updated documentation)
  * bug [#123] [ContentBundle][Bridge] Article's body is not pre-filled (takeit)
  * bug [#122] [Templates System] add custom cache key generator for meta objects
  * feature [#120] Make routes of "collection" type accessible (takeit)